Address PR review: extract shared finetune utils, harden test cleanup - README updated to add dataset details - Extract formatProgress, createFilteredLogger into examples/utils.js,   removing ~600 lines of duplicated code across 5 finetune examples - Restructure smart-home-finetuned-test.js cleanup: single outer finally   for loader.close(), inner try/catch for each client.unload(), no more   process.exit() in baseline catch that skipped loader cleanup - Add early guard for missing LoRA adapter in test script with clear   error message pointing to the finetune script - Move generated report from examples/input/ to the adapter output   directory (./smart-home-lora/) so it doesn't pollute source-controlled   training data - Add dataset diversity note in finetune script and README explaining   intentional cross-domain training for behavioral pattern transfer - Normalize require('../../index.js') to require('../../index') for   consistency with all other examples
