+++ packages/qvac-lib-infer-llamacpp-llm/examples/finetune/showcase/smart-home-finetune.js
@@ -3,92 +3,26 @@
 const LlamaClient = require('../../../index')
 const FilesystemDL = require('@qvac/dl-filesystem')
 const process = require('bare-process')
-const { downloadModel } = require('../../utils')
+const { downloadModel, formatProgress, createFilteredLogger } = require('../../utils')
 
 const MODEL = {
   name: 'Qwen3-0.6B-Q8_0.gguf',
   url: 'https://huggingface.co/unsloth/Qwen3-0.6B-GGUF/resolve/main/Qwen3-0.6B-Q8_0.gguf'
 }
 
+// The dataset intentionally covers diverse tool domains (medical, irrigation,
+// quantum, etc.) — not just the 4 smart-home tools used in evaluation. The goal
+// is to teach the model the *behavioral pattern* (user request -> short think ->
+// structured <tool_call> output) rather than memorize specific tool names.
 const TRAIN_DATASET = './examples/input/smart_home_specialist_train.jsonl'
 const CHECKPOINT_DIR = './smart-home-lora-checkpoints'
 const OUTPUT_DIR = './smart-home-lora'
-
-function formatTime (ms) {
-  if (!Number.isFinite(ms) || ms < 0) return '--:--'
-  const totalSec = Math.floor(ms / 1000)
-  const h = Math.floor(totalSec / 3600)
-  const m = Math.floor((totalSec % 3600) / 60)
-  const s = totalSec % 60
-  if (h > 0) return `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`
-  return `${m}:${String(s).padStart(2, '0')}`
-}
-
-function makeProgressBar (current, total, width) {
-  width = width || 30
-  if (!total || total <= 0) return '[' + ' '.repeat(width) + ']'
-  const filled = Math.round((current / total) * width)
-  return '[' + '\u2588'.repeat(filled) + '\u2591'.repeat(width - filled) + ']'
-}
-
-function formatProgress (stats, totalEpochs) {
-  const isTrain = stats.is_train !== false
-  const phase = isTrain ? 'train' : 'val  '
-  const epoch = Number.isFinite(stats.current_epoch) ? stats.current_epoch + 1 : 1
-  const bar = makeProgressBar(stats.current_batch, stats.total_batches)
-  const batchStr = `${stats.current_batch}/${stats.total_batches}`
-  const loss = Number.isFinite(stats.loss) ? stats.loss.toFixed(4) : 'n/a'
-  const acc = Number.isFinite(stats.accuracy) ? (stats.accuracy * 100).toFixed(1) + '%' : 'n/a'
-  const elapsed = formatTime(stats.elapsed_ms)
-  const eta = formatTime(stats.eta_ms)
-  const stepStr = isTrain ? ` step=${stats.global_steps}` : ''
-  return `${phase} epoch ${epoch}/${totalEpochs} ${bar} ${batchStr} | loss=${loss} acc=${acc}${stepStr} | ${elapsed}<${eta}`
-}
 
 async function main () {
   let client
   let loader
 
-  const originalConsoleLog = console.log
-  const originalConsoleInfo = console.info
-  const originalConsoleWarn = console.warn
-
-  const shouldSuppressMessage = (args) => {
-    const message = args.join(' ')
-    return message && message.includes('No response found for job')
-  }
-
-  console.log = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleLog.apply(console, args)
-  }
-
-  console.info = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleInfo.apply(console, args)
-  }
-
-  console.warn = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleWarn.apply(console, args)
-  }
-
-  const filteredLogger = {
-    info: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleInfo.apply(console, args)
-    },
-    log: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleLog.apply(console, args)
-    },
-    warn: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleWarn.apply(console, args)
-    },
-    error: console.error.bind(console),
-    debug: console.debug.bind(console)
-  }
+  const { logger: filteredLogger, restore: restoreConsole } = createFilteredLogger()
 
   try {
     console.log('='.repeat(70))
@@ -176,9 +110,7 @@ async function main () {
     console.error('Stack:', error.stack)
     process.exit(1)
   } finally {
-    console.log = originalConsoleLog
-    console.info = originalConsoleInfo
-    console.warn = originalConsoleWarn
+    restoreConsole()
 
     if (client) {
       try {
