+++ packages/qvac-lib-infer-llamacpp-llm/examples/finetune/simple-lora-finetune-pause-inference-resume.js
@@ -5,7 +5,7 @@ const FilesystemDL = require('@qvac/dl-filesystem')
 const process = require('bare-process')
 const path = require('bare-path')
 const fs = require('bare-fs')
-const https = require('bare-https')
+const { downloadModel, formatProgress, createFilteredLogger } = require('../utils')
 
 const MODEL = {
   name: 'Qwen3-0.6B-Q8_0.gguf',
@@ -32,129 +32,6 @@ function waitForProgress (handle, minSteps, timeoutMs) {
     }
     handle.on('stats', onStats)
   })
-}
-
-function formatTime (ms) {
-  if (!Number.isFinite(ms) || ms < 0) return '--:--'
-  const totalSec = Math.floor(ms / 1000)
-  const h = Math.floor(totalSec / 3600)
-  const m = Math.floor((totalSec % 3600) / 60)
-  const s = totalSec % 60
-  if (h > 0) return `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`
-  return `${m}:${String(s).padStart(2, '0')}`
-}
-
-function makeProgressBar (current, total, width) {
-  width = width || 20
-  if (!total || total <= 0) return '[' + ' '.repeat(width) + ']'
-  const filled = Math.round((current / total) * width)
-  return '[' + '\u2588'.repeat(filled) + '\u2591'.repeat(width - filled) + ']'
-}
-
-function formatProgress (stats, totalEpochs) {
-  const isTrain = stats.is_train !== false
-  const phase = isTrain ? 'train' : 'val  '
-  const epoch = Number.isFinite(stats.current_epoch) ? stats.current_epoch + 1 : 1
-  const bar = makeProgressBar(stats.current_batch, stats.total_batches)
-  const batchStr = `${stats.current_batch}/${stats.total_batches}`
-  const loss = Number.isFinite(stats.loss) ? stats.loss.toFixed(4) : 'n/a'
-  const acc = Number.isFinite(stats.accuracy) ? (stats.accuracy * 100).toFixed(1) + '%' : 'n/a'
-  const elapsed = formatTime(stats.elapsed_ms)
-  const eta = formatTime(stats.eta_ms)
-  const stepStr = isTrain ? ` step=${stats.global_steps}` : ''
-  return `${phase} epoch ${epoch}/${totalEpochs} ${bar} ${batchStr} | loss=${loss} acc=${acc}${stepStr} | ${elapsed}<${eta}`
-}
-
-async function downloadFile (url, dest) {
-  return new Promise((resolve, reject) => {
-    let resolved = false
-    const safeResolve = () => {
-      if (!resolved) {
-        resolved = true
-        resolve()
-      }
-    }
-    const safeReject = (err) => {
-      if (!resolved) {
-        resolved = true
-        reject(err)
-      }
-    }
-
-    const file = fs.createWriteStream(dest)
-
-    file.on('error', (err) => {
-      file.destroy()
-      fs.unlink(dest, () => safeReject(err))
-    })
-
-    const req = https.request(url, response => {
-      if ([301, 302, 307, 308].includes(response.statusCode)) {
-        file.destroy()
-        fs.unlink(dest, (unlinkErr) => {
-          if (unlinkErr && unlinkErr.code !== 'ENOENT') {
-            return safeReject(unlinkErr)
-          }
-
-          let redirectUrl = response.headers.location
-          if (redirectUrl.startsWith('/')) {
-            const originalUrl = new URL(url)
-            redirectUrl = `${originalUrl.protocol}//${originalUrl.host}${redirectUrl}`
-          }
-
-          downloadFile(redirectUrl, dest)
-            .then(safeResolve)
-            .catch(safeReject)
-        })
-        return
-      }
-
-      if (response.statusCode !== 200) {
-        file.destroy()
-        fs.unlink(dest, () => safeReject(new Error(`Download failed: HTTP ${response.statusCode} from ${url}`)))
-        return
-      }
-
-      response.on('error', (err) => {
-        file.destroy()
-        fs.unlink(dest, () => safeReject(err))
-      })
-
-      response.pipe(file)
-
-      file.on('close', () => {
-        safeResolve()
-      })
-    })
-
-    req.on('error', err => {
-      file.destroy()
-      fs.unlink(dest, () => safeReject(err))
-    })
-
-    req.end()
-  })
-}
-
-async function ensureModel ({ modelName, downloadUrl }) {
-  const modelDir = path.resolve('./models')
-
-  const modelPath = path.join(modelDir, modelName)
-
-  if (fs.existsSync(modelPath)) {
-    const stats = fs.statSync(modelPath)
-    console.log(`Found ${modelName}: ${(stats.size / 1024 / 1024).toFixed(1)}MB`)
-    return [modelName, modelDir]
-  }
-
-  fs.mkdirSync(modelDir, { recursive: true })
-  console.log(`Downloading test model ${modelName}...`)
-
-  await downloadFile(downloadUrl, modelPath)
-
-  const stats = fs.statSync(modelPath)
-  console.log(`Model ready: ${(stats.size / 1024 / 1024).toFixed(1)}MB`)
-  return [modelName, modelDir]
 }
 
 function findPauseCheckpoint (checkpointDir) {
@@ -194,56 +71,14 @@ async function runInference (client, description, messages) {
 }
 
 async function main () {
-  const [modelName, modelDir] = await ensureModel({
-    modelName: MODEL.name,
-    downloadUrl: MODEL.url
-  })
+  const [modelName, modelDir] = await downloadModel(MODEL.url, MODEL.name)
 
   const trainDatasetPath = './examples/input/small_train_HF.jsonl'
   const evalDatasetPath = './examples/input/small_eval_HF.jsonl'
 
   const loader = new FilesystemDL({ dirPath: modelDir })
 
-  const originalConsoleLog = console.log
-  const originalConsoleInfo = console.info
-  const originalConsoleWarn = console.warn
-
-  const shouldSuppressMessage = (args) => {
-    const message = args.join(' ')
-    return message && message.includes('No response found for job')
-  }
-
-  console.log = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleLog.apply(console, args)
-  }
-
-  console.info = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleInfo.apply(console, args)
-  }
-
-  console.warn = (...args) => {
-    if (shouldSuppressMessage(args)) return
-    originalConsoleWarn.apply(console, args)
-  }
-
-  const filteredLogger = {
-    info: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleInfo.apply(console, args)
-    },
-    log: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleLog.apply(console, args)
-    },
-    warn: (...args) => {
-      if (shouldSuppressMessage(args)) return
-      originalConsoleWarn.apply(console, args)
-    },
-    error: console.error.bind(console),
-    debug: console.debug.bind(console)
-  }
+  const { logger: filteredLogger, restore: restoreConsole } = createFilteredLogger()
 
   const args = {
     loader,
@@ -465,9 +300,7 @@ async function main () {
     console.error('Stack:', error.stack)
     process.exit(1)
   } finally {
-    console.log = originalConsoleLog
-    console.info = originalConsoleInfo
-    console.warn = originalConsoleWarn
+    restoreConsole()
 
     if (client) {
       try {
