+++ packages/qvac-lib-infer-llamacpp-llm/examples/utils.js
@@ -72,4 +72,86 @@ async function downloadModel (url, filename) {
   })
 }
 
-module.exports = { downloadModel }
+function formatTime (ms) {
+  if (!Number.isFinite(ms) || ms < 0) return '--:--'
+  const totalSec = Math.floor(ms / 1000)
+  const h = Math.floor(totalSec / 3600)
+  const m = Math.floor((totalSec % 3600) / 60)
+  const s = totalSec % 60
+  if (h > 0) return `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`
+  return `${m}:${String(s).padStart(2, '0')}`
+}
+
+function makeProgressBar (current, total, width) {
+  width = width || 20
+  if (!total || total <= 0) return '[' + ' '.repeat(width) + ']'
+  const filled = Math.round((current / total) * width)
+  return '[' + '\u2588'.repeat(filled) + '\u2591'.repeat(width - filled) + ']'
+}
+
+function formatProgress (stats, totalEpochs) {
+  const isTrain = stats.is_train !== false
+  const phase = isTrain ? 'train' : 'val  '
+  const epoch = Number.isFinite(stats.current_epoch) ? stats.current_epoch + 1 : 1
+  const bar = makeProgressBar(stats.current_batch, stats.total_batches)
+  const batchStr = `${stats.current_batch}/${stats.total_batches}`
+  const loss = Number.isFinite(stats.loss) ? stats.loss.toFixed(4) : 'n/a'
+  const acc = Number.isFinite(stats.accuracy) ? (stats.accuracy * 100).toFixed(1) + '%' : 'n/a'
+  const elapsed = formatTime(stats.elapsed_ms)
+  const eta = formatTime(stats.eta_ms)
+  const stepStr = isTrain ? ` step=${stats.global_steps}` : ''
+  return `${phase} epoch ${epoch}/${totalEpochs} ${bar} ${batchStr} | loss=${loss} acc=${acc}${stepStr} | ${elapsed}<${eta}`
+}
+
+function createFilteredLogger () {
+  const originalConsoleLog = console.log
+  const originalConsoleInfo = console.info
+  const originalConsoleWarn = console.warn
+
+  const shouldSuppress = (args) => {
+    const message = args.join(' ')
+    return message && message.includes('No response found for job')
+  }
+
+  console.log = (...args) => {
+    if (shouldSuppress(args)) return
+    originalConsoleLog.apply(console, args)
+  }
+
+  console.info = (...args) => {
+    if (shouldSuppress(args)) return
+    originalConsoleInfo.apply(console, args)
+  }
+
+  console.warn = (...args) => {
+    if (shouldSuppress(args)) return
+    originalConsoleWarn.apply(console, args)
+  }
+
+  const logger = {
+    info: (...args) => {
+      if (shouldSuppress(args)) return
+      originalConsoleInfo.apply(console, args)
+    },
+    log: (...args) => {
+      if (shouldSuppress(args)) return
+      originalConsoleLog.apply(console, args)
+    },
+    warn: (...args) => {
+      if (shouldSuppress(args)) return
+      originalConsoleWarn.apply(console, args)
+    },
+    error: console.error.bind(console),
+    debug: console.debug.bind(console)
+  }
+
+  function restore () {
+    console.log = originalConsoleLog
+    console.info = originalConsoleInfo
+    console.warn = originalConsoleWarn
+  }
+
+  return { logger, restore }
+}
+
+module.exports = { downloadModel, formatProgress, createFilteredLogger }
